Basic PNG parsing, Going to take a break for a while and work on a more.. 'realistic' opengl project and come back to this once I have a deeper understanding

+++ inc/Engine/factory.hpp
@@ -11,11 +11,11 @@ constexpr Vec3 defaultScale{1.0f, 1.0f, 1.0f};
 bool createTriangle(Mesh& mesh, const Vec2& size = defaultSize2D);
 
 bool createSquare(Mesh& mesh, const Vec2& size = defaultSize2D);
-bool createSquareGrid(Model* models, int startIndex, int count, const Vec2& spacing, const Vec3& rotation = defaultRotation, const Vec2& size = defaultSize2D);
+bool createSquareGrid(Model* models, size_t startIndex, int count, const Vec2& spacing, const Vec3& rotation = defaultRotation, const Vec2& size = defaultSize2D);
 
 bool createCube(Mesh& mesh, const Vec3& size = defaultSize3D);
-bool createCubeGrid(Model* models, int startIndex, int count, const Vec2& spacing, const Vec3& rotation = defaultRotation, const Vec3& size = defaultSize3D);
+bool createCubeGrid(Model* models, size_t startIndex, int count, const Vec2& spacing, const Vec3& rotation = defaultRotation, const Vec3& size = defaultSize3D);
 
 // Meshes
 bool createMeshPath(Mesh* mesh, const char* path, const char* type, const Vec3& scale = defaultScale, const bool hasNormals = false);
-bool createModelGrid(Model* models, const char* path, int startIndex, int count, const Vec2& spacing, const Vec3& rotation = defaultRotation, const Vec3& scale = defaultScale, const bool hasNormals = false);
+bool createModelGrid(Model* models, const char* path, size_t startIndex, int count, const Vec2& spacing, const Vec3& rotation = defaultRotation, const Vec3& scale = defaultScale, const bool hasNormals = false);
